refactor: extract auth and home data logic into custom hooks
